refactor: 불필요한 파일 삭제 및 Router 리팩토링
- 사용하지 않는 index.ts 파일 삭제
- oauth.ts 파일을 상위(router) 디렉토리로 이동
- api들을 관리하는 api router 생성 및 api 호출 url에 /api를 추가
- router naming 통일
- api 호출 시 jwt 인증이 필요하도록 middleware 추가
