+++ server/src/router/chatroom-router.ts
@@ -1,9 +1,8 @@
 import express from 'express';
-import passport from 'passport';
 import ChatroomController from '@controller/chatroom-controller';
 
 const router = express.Router();
 
-router.post('/', passport.authenticate('jwt'), ChatroomController.createChatroom);
+router.post('/', ChatroomController.createChatroom);
 
 export default router;
