+++ server/src/router/index.ts
@@ -1,18 +1,13 @@
 import express from 'express';
-import oauthRouter from '@router/api/oauth';
-import UserRouter from '@router/user-router';
-import ChatroomRouter from '@router/chatroom-router';
-import userChatroomRouter from '@router/user-chatroom-router';
+import passport from 'passport';
+import oauthRouter from '@router/oauth';
+import apiRouter from '@router/api';
 
 const router = express.Router();
 
-router.get('/api', (req, res) => {
-  res.send('router test');
-});
-
 router.use('/oauth', oauthRouter);
-router.use('/users', UserRouter);
-router.use('/chatrooms', ChatroomRouter);
-router.use('/user-chatrooms', userChatroomRouter);
+
+router.use(passport.authenticate('jwt'));
+router.use('/api', apiRouter);
 
 export default router;
