+++ server/src/router/user-chatroom-router.ts
@@ -1,9 +1,8 @@
 import express from 'express';
-import passport from 'passport';
 import UserChatroomController from '../controller/user-chatroom-controller';
 
 const router = express.Router();
 
-router.get('/', passport.authenticate('jwt'), UserChatroomController.getUserChatrooms);
+router.get('/', UserChatroomController.getUserChatrooms);
 
 export default router;
